--- servicecatalog_puppet/commands/bootstrap.py
@@ -1,78 +1,0 @@
-from threading import Thread
-import os
-import click
-from betterboto import client as betterboto_client
-from jinja2 import Template
-
-from servicecatalog_puppet.asset_helpers import read_from_site_packages
-from servicecatalog_puppet.constants import BOOTSTRAP_STACK_NAME, SERVICE_CATALOG_PUPPET_REPO_NAME
-from servicecatalog_puppet.core import get_regions, get_org_iam_role_arn
-
-
-def do_bootstrap(puppet_version):
-    click.echo('Starting bootstrap')
-    ALL_REGIONS = get_regions(os.environ.get("AWS_DEFAULT_REGION"))
-    with betterboto_client.MultiRegionClientContextManager('cloudformation', ALL_REGIONS) as clients:
-        click.echo('Creating {}-regional'.format(BOOTSTRAP_STACK_NAME))
-        threads = []
-        template = read_from_site_packages('{}.template.yaml'.format('{}-regional'.format(BOOTSTRAP_STACK_NAME)))
-        template = Template(template).render(VERSION=puppet_version)
-        args = {
-            'StackName': '{}-regional'.format(BOOTSTRAP_STACK_NAME),
-            'TemplateBody': template,
-            'Capabilities': ['CAPABILITY_IAM'],
-            'Parameters': [
-                {
-                    'ParameterKey': 'Version',
-                    'ParameterValue': puppet_version,
-                    'UsePreviousValue': False,
-                },
-                {
-                    'ParameterKey': 'DefaultRegionValue',
-                    'ParameterValue': os.environ.get('AWS_DEFAULT_REGION'),
-                    'UsePreviousValue': False,
-                },
-            ],
-        }
-        for client_region, client in clients.items():
-            process = Thread(name=client_region, target=client.create_or_update, kwargs=args)
-            process.start()
-            threads.append(process)
-        for process in threads:
-            process.join()
-        click.echo('Finished creating {}-regional'.format(BOOTSTRAP_STACK_NAME))
-
-    with betterboto_client.ClientContextManager('cloudformation') as cloudformation:
-        click.echo('Creating {}'.format(BOOTSTRAP_STACK_NAME))
-        template = read_from_site_packages('{}.template.yaml'.format(BOOTSTRAP_STACK_NAME))
-        template = Template(template).render(VERSION=puppet_version, ALL_REGIONS=ALL_REGIONS)
-        args = {
-            'StackName': BOOTSTRAP_STACK_NAME,
-            'TemplateBody': template,
-            'Capabilities': ['CAPABILITY_NAMED_IAM'],
-            'Parameters': [
-                {
-                    'ParameterKey': 'Version',
-                    'ParameterValue': puppet_version,
-                    'UsePreviousValue': False,
-                },
-                {
-                    'ParameterKey': 'OrgIamRoleArn',
-                    'ParameterValue': str(get_org_iam_role_arn()),
-                    'UsePreviousValue': False,
-                },
-            ],
-        }
-        cloudformation.create_or_update(**args)
-
-    click.echo('Finished creating {}.'.format(BOOTSTRAP_STACK_NAME))
-    with betterboto_client.ClientContextManager('codecommit') as codecommit:
-        response = codecommit.get_repository(repositoryName=SERVICE_CATALOG_PUPPET_REPO_NAME)
-        clone_url = response.get('repositoryMetadata').get('cloneUrlHttp')
-        clone_command = "git clone --config 'credential.helper=!aws codecommit credential-helper $@' " \
-                        "--config 'credential.UseHttpPath=true' {}".format(clone_url)
-        click.echo(
-            'You need to clone your newly created repo now and will then need to seed it: \n{}'.format(
-                clone_command
-            )
-        )
