+++ servicecatalog_puppet/core.py
@@ -13,163 +13,3 @@ from servicecatalog_puppet.constants import HOME_REGION_PARAM_NAME, CONFIG_PARAM
 
 logger = logging.getLogger()
 
-
-def get_regions(default_region=None):
-    logger.info("getting regions,  default_region: {}".format(default_region))
-    with betterboto_client.ClientContextManager(
-            'ssm',
-            region_name=default_region if default_region else get_home_region()
-    ) as ssm:
-        response = ssm.get_parameter(Name=CONFIG_PARAM_NAME)
-        config = yaml.safe_load(response.get('Parameter').get('Value'))
-        return config.get('regions')
-
-
-def get_home_region():
-    with betterboto_client.ClientContextManager('ssm') as ssm:
-        response = ssm.get_parameter(Name=HOME_REGION_PARAM_NAME)
-        return response.get('Parameter').get('Value')
-
-
-def get_org_iam_role_arn():
-    with betterboto_client.ClientContextManager('ssm', region_name=get_home_region()) as ssm:
-        try:
-            response = ssm.get_parameter(Name=CONFIG_PARAM_NAME_ORG_IAM_ROLE_ARN)
-            return yaml.safe_load(response.get('Parameter').get('Value'))
-        except ssm.exceptions.ParameterNotFound as e:
-            logger.info("No parameter set for: {}".format(CONFIG_PARAM_NAME_ORG_IAM_ROLE_ARN))
-            return None
-
-
-def get_provisioning_artifact_id_for(portfolio_name, product_name, version_name, account_id, region):
-    logger.info("Getting provisioning artifact id for: {} {} {} in the region: {} of account: {}".format(
-        portfolio_name, product_name, version_name, region, account_id
-    ))
-    role = "arn:aws:iam::{}:role/{}".format(account_id, 'servicecatalog-puppet/PuppetRole')
-    with betterboto_client.CrossAccountClientContextManager(
-            'servicecatalog', role, "-".join([account_id, region]), region_name=region
-    ) as cross_account_servicecatalog:
-        product_id = None
-        version_id = None
-        portfolio_id = None
-        args = {}
-        while True:
-            response = cross_account_servicecatalog.list_accepted_portfolio_shares()
-            assert response.get('NextPageToken') is None, "Pagination not supported"
-            for portfolio_detail in response.get('PortfolioDetails'):
-                if portfolio_detail.get('DisplayName') == portfolio_name:
-                    portfolio_id = portfolio_detail.get('Id')
-                    break
-
-            if portfolio_id is None:
-                response = cross_account_servicecatalog.list_portfolios()
-                for portfolio_detail in response.get('PortfolioDetails', []):
-                    if portfolio_detail.get('DisplayName') == portfolio_name:
-                        portfolio_id = portfolio_detail.get('Id')
-                        break
-
-            assert portfolio_id is not None, "Could not find portfolio"
-            logger.info("Found portfolio: {}".format(portfolio_id))
-
-            args['PortfolioId'] = portfolio_id
-            response = cross_account_servicecatalog.search_products_as_admin(
-                **args
-            )
-            for product_view_details in response.get('ProductViewDetails'):
-                product_view = product_view_details.get('ProductViewSummary')
-                if product_view.get('Name') == product_name:
-                    logger.info('Found product: {}'.format(product_view))
-                    product_id = product_view.get('ProductId')
-            if response.get('NextPageToken', None) is not None:
-                args['PageToken'] = response.get('NextPageToken')
-            else:
-                break
-        assert product_id is not None, "Did not find product looking for"
-
-        response = cross_account_servicecatalog.list_provisioning_artifacts(
-            ProductId=product_id
-        )
-        assert response.get('NextPageToken') is None, "Pagination not support"
-        for provisioning_artifact_detail in response.get('ProvisioningArtifactDetails'):
-            if provisioning_artifact_detail.get('Name') == version_name:
-                version_id = provisioning_artifact_detail.get('Id')
-        assert version_id is not None, "Did not find version looking for"
-        return product_id, version_id
-
-
-def generate_bucket_policies_for_shares(deployment_map, puppet_account_id):
-    shares = {
-        'accounts': [],
-        'organizations': [],
-    }
-    for account_id, deployment in deployment_map.items():
-        if account_id == puppet_account_id:
-            continue
-        if deployment.get('expanded_from') is None:
-            if account_id not in shares['accounts']:
-                shares['accounts'].append(account_id)
-        else:
-            if deployment.get('organization') not in shares['organizations']:
-                shares['organizations'].append(deployment.get('organization'))
-    return shares
-
-
-def write_share_template(portfolio_use_by_account, region, host_account_id, sharing_policies):
-    output = os.path.sep.join([TEMPLATES, 'shares', region])
-    if not os.path.exists(output):
-        os.makedirs(output)
-
-    with open(os.sep.join([output, "shares.template.yaml"]), 'w') as f:
-        f.write(
-            env.get_template('shares.template.yaml.j2').render(
-                portfolio_use_by_account=portfolio_use_by_account,
-                host_account_id=host_account_id,
-                HOME_REGION=get_home_region(),
-                sharing_policies=sharing_policies,
-            )
-        )
-
-
-def create_share_template(deployment_map, puppet_account_id):
-    logger.info("deployment_map: {}".format(deployment_map))
-    ALL_REGIONS = get_regions()
-    for region in ALL_REGIONS:
-        logger.info("starting to build shares for region: {}".format(region))
-        with betterboto_client.ClientContextManager('servicecatalog', region_name=region) as servicecatalog:
-            portfolio_ids = {}
-            args = {}
-            while True:
-
-                response = servicecatalog.list_portfolios(
-                    **args
-                )
-
-                for portfolio_detail in response.get('PortfolioDetails'):
-                    portfolio_ids[portfolio_detail.get('DisplayName')] = portfolio_detail.get('Id')
-
-                if response.get('PageToken') is not None:
-                    args['PageToken'] = response.get('PageToken')
-                else:
-                    break
-
-            logger.info("Portfolios in use in region: {}".format(portfolio_ids))
-
-            portfolio_use_by_account = {}
-            for account_id, launch_details in deployment_map.items():
-                if portfolio_use_by_account.get(account_id) is None:
-                    portfolio_use_by_account[account_id] = []
-                for launch_id, launch in launch_details.get('launches').items():
-                    logger.info("portfolio ids: {}".format(portfolio_ids))
-                    p = portfolio_ids[launch.get('portfolio')]
-                    if p not in portfolio_use_by_account[account_id]:
-                        portfolio_use_by_account[account_id].append(p)
-            host_account_id = response.get('PortfolioDetails')[0].get('ARN').split(":")[4]
-            sharing_policies = generate_bucket_policies_for_shares(deployment_map, puppet_account_id)
-            write_share_template(portfolio_use_by_account, region, host_account_id, sharing_policies)
-
-
-template_dir = resolve_from_site_packages('templates')
-env = Environment(
-    loader=FileSystemLoader(template_dir),
-    extensions=['jinja2.ext.do'],
-)
