+++ servicecatalog_puppet/manifest_utils.py
@@ -1,5 +1,10 @@
 import yaml
 import logging
+from copy import deepcopy
+
+from cli_command_helpers import expand_path, expand_ou
+from cli_commands import logger
+from macros import macros
 
 logger = logging.getLogger(__file__)
 
@@ -131,3 +136,65 @@ def check_for_duplicate_products_in_launches(launches_by_product):
                         )
                     )
     logger.info('Finished checking for duplicate products by account listed twice')
+
+
+def expand_manifest(manifest, client):
+    new_manifest = deepcopy(manifest)
+    new_accounts = new_manifest['accounts'] = []
+
+    logger.info('Starting the expand')
+
+    for account in manifest.get('accounts'):
+        if account.get('account_id'):
+            logger.info("Found an account: {}".format(account.get('account_id')))
+            new_accounts.append(account)
+        elif account.get('ou'):
+            ou = account.get('ou')
+            logger.info("Found an ou: {}".format(ou))
+            if ou.startswith('/'):
+                new_accounts += expand_path(account, client)
+            else:
+                new_accounts += expand_ou(account, client)
+
+    logger.debug(new_accounts)
+
+    for parameter_name, parameter_details in new_manifest.get('parameters', {}).items():
+        if parameter_details.get('macro'):
+            macro_to_run = macros.get(parameter_details.get('macro').get('method'))
+            result = macro_to_run(client, parameter_details.get('macro').get('args'))
+            parameter_details['default'] = result
+            del parameter_details['macro']
+
+    for first_account in new_accounts:
+        for parameter_name, parameter_details in first_account.get('parameters', {}).items():
+            if parameter_details.get('macro'):
+                macro_to_run = macros.get(parameter_details.get('macro').get('method'))
+                result = macro_to_run(client, parameter_details.get('macro').get('args'))
+                parameter_details['default'] = result
+                del parameter_details['macro']
+
+        times_seen = 0
+        for second_account in new_accounts:
+            if first_account.get('account_id') == second_account.get('account_id'):
+                times_seen += 1
+                if times_seen > 1:
+                    message = "{} has been seen twice.".format(first_account.get('account_id'))
+                    if first_account.get('expanded_from'):
+                        message += "  It was included due to it being in the ou: {}".format(
+                            first_account.get('expanded_from')
+                        )
+                    if second_account.get('expanded_from'):
+                        message += "  It was included due to it being in the ou: {}".format(
+                            second_account.get('expanded_from')
+                        )
+                    raise Exception(message)
+
+    for launch_name, launch_details in new_manifest.get('launches').items():
+        for parameter_name, parameter_details in launch_details.get('parameters', {}).items():
+            if parameter_details.get('macro'):
+                macro_to_run = macros.get(parameter_details.get('macro').get('method'))
+                result = macro_to_run(client, parameter_details.get('macro').get('args'))
+                parameter_details['default'] = result
+                del parameter_details['macro']
+
+    return new_manifest
